+++ projects/ngrx-rtk-query/src/lib/provide-store-api.ts
@@ -1,7 +1,12 @@
-import { ENVIRONMENT_INITIALIZER, inject, makeEnvironmentProviders, type EnvironmentProviders } from '@angular/core';
+import {
+  ENVIRONMENT_INITIALIZER,
+  Injector,
+  inject,
+  makeEnvironmentProviders,
+  type EnvironmentProviders,
+} from '@angular/core';
 import { provideState } from '@ngrx/store';
 import { setupListeners as setupListenersFn, type Api } from '@reduxjs/toolkit/query';
-import { ThunkService } from './thunk.service';
 
 export interface StoreQueryConfig {
   setupListeners?: Parameters<typeof setupListenersFn>[1] | false;
@@ -18,7 +23,8 @@ export function provideStoreApi(
       provide: ENVIRONMENT_INITIALIZER,
       multi: true,
       useValue() {
-        inject(ThunkService).init();
+        const injector = inject(Injector);
+        Object.assign(api, { injector });
       },
     },
     provideState(api.reducerPath, api.reducer),