+++ projects/ngrx-rtk-query/src/public-api.ts
@@ -2,7 +2,7 @@ import { angularHooksModule, angularHooksModuleName } from './lib/module';
 
 export * from '@reduxjs/toolkit/query';
 export { fetchBaseQuery } from './lib/fetch-base-query';
-export { provideStoreApi } from './lib/provide-rtk-query';
+export { provideStoreApi } from './lib/provide-store-api';
 
 export { createApi } from './lib/create-api';